// src/types/SkillTypes.ts
export type SkillType = {
  id: number;
  name: string;
  subName: string;
  slotLv: string;
  effect: string;
  levell: {
    one?: string;
    two?: string;
    three?: string;
    four?: string;
    five?: string;
    six?: string;
    seven?: string;
  };
  explanation: {
    one?: string;
    two?: string;
    three?: string;
    four?: string;
    five?: string;
    six?: string;
    seven?: string;
  };
};